% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-census.R
\name{acs.fetch.and.cache}
\alias{acs.fetch.and.cache}
\title{Download census data, employing caching to avoid repeated downloads.}
\usage{
acs.fetch.and.cache(...)
}
\description{
A thin caching wrapper around \code{acs.fetch(...)} in the acs package.
}
\examples{
  acs.fetch.and.cache(...)
}

