% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basic.R
\name{basicFeatures}
\alias{basicFeatures}
\title{Implementation of several basic statistical features}
\usage{
basicFeatures(meterData, ...)
}
\arguments{
\item{meterData}{An instance of \code{\link{MeterDataClass}} that provides the data used for all the implemented feature calculations.}
}
\description{
Function that implements a set of common statistical features that can be run on an instance of \code{\link{MeterDataClass}}.
}
\details{
\code{basicFeatures} is called by passing in an instance of a MeterDataClass, which provides the meter data in both linear series 
  and day-per-row matrix formats. These data structures and some associated indices and intermediate data structures are
  re-used throughout this method, which tries to take advantage of the performance benefits of such re-use.
  This function is often called in the context of the \code{iterator.*} suite of functions, which can loop through large sets of 
  meters, calling feature functions on each.
}
\examples{
\dontrun{
DATA_SOURCE = TestData(n=10)
meterData = DATA_SOURCE$getMeterDataClass(DATA_SOURCE$getIds()[1])
features = basicFeatures(meterData)
names(features)
class(features)
}
}

