% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetDocs.R
\docType{data}
\name{CA_ZIP_CLIMATE}
\alias{CA_ZIP_CLIMATE}
\title{A zip code to CA climate zone mapping}
\format{A data frame with 1706 rows and 2 variables:
\describe{
  \item{ZIP.Code}{zip code, as a number}
  \item{Building.Climate.Zone}{the CA CEC climate zone the zip code is in}
}}
\source{
\url{http://CEC.web.site.somewhere/}
}
\usage{
data(CA_ZIP_CLIMATE)
}
\description{
A dataset containing columns with US zip codes and the corresponding CA CEC climate zones
}
\keyword{datasets}

