% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetDocs.R
\docType{data}
\name{CENSUS_GAZ}
\alias{CENSUS_GAZ}
\title{A summary of census statistics for each ZCTA in the census}
\format{A data frame with 33120 rows and 9 variables:
\describe{
  \item{ZCTA}{zip code tabulation area as a 0 padded string}
  \item{POP10}{Population in 2010}
  \item{ALAND}{Land area in sqft?}
  \item{AWATER}{Water area in sqft?}
  \item{ALAND_SQMI}{Land area in sqare miles}
  \item{AWATER_SQMI}{Water area in square miles}
  \item{INTPTLAT}{Latitude of the area}
  \item{INTPTLONG}{Longitude of the area}
}}
\source{
\url{http://us.census.link/}
}
\usage{
data(CENSUS_GAZ)
}
\description{
A dataset containing summary census statistics for each ZCTA in the census
}
\keyword{datasets}

