% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{cleanFeatureDF}
\alias{cleanFeatureDF}
\title{Clean up feature data in preparation for saving}
\usage{
cleanFeatureDF(features)
}
\arguments{
\item{features}{The data frame of feature to be cleaned up}
}
\value{
A copy of the original data frame that is cleaned up
}
\description{
This function renames data columns for export via fixNames(), converts factors 
into characters, and checks for id and zip5 colums
}

