% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-dbUtil.R
\name{clearCons}
\alias{clearCons}
\title{Close and clear all connections}
\usage{
clearCons(cfg)
}
\arguments{
\item{cfg}{a named list of configuration options, often loaded from a flat configuration file using \code{\link{dbCfg}}.}
}
\description{
Utility function to close/clear all active db connections, which can sometimes get left open, especially when errors
prevent connections from being closed.
}

