% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-tree.R
\name{ctree.boot}
\alias{ctree.boot}
\title{bootstrap several runs of the ctree algorithm}
\usage{
ctree.boot(df, nRun = 100, nSample = NULL, ctl = NULL, responseVar,
  ignoreCols)
}
\arguments{
\item{df}{data.frame with response variable to be explained and all explanatory variables for the ctree to run with}

\item{nRun}{number of bootstrapping runs to perform}

\item{nSample}{number of samples (with replacement) for each run}

\item{ctl}{ctree control instance to be passed to ctree algorithm to control stopping criteria, etc.}

\item{responseVar}{the name of the variable that the ctree is trying to explain}

\item{ignoreCols}{an optional list of columns in df, but that should be excluded from the ctree modeling}
}
\description{
This function repeatedly runs the ctree algorithm on random sub-samples of the underlying data
to provide a measure of robustness for the individual ctree results.
}

