% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-tree.R
\name{ctree.run}
\alias{ctree.run}
\title{run a ctree model}
\usage{
ctree.run(df, fmla = NULL, sub = T, ctl = NULL, responseVar, ignoreCols,
  ...)
}
\arguments{
\item{df}{data.frame with response variable to be explained and all explanatory variables to be used in ctree classificaiton}

\item{fmla}{formula, referring to values in the df, specifying the variables to be used to train the ctree model}

\item{ctl}{partykit::ctree_control instance to specify ctree model parameters}

\item{responseVar}{the name of the variable that the ctree is trying to explain}

\item{ignoreCols}{an optional list of columns in df, but that should be excluded from the ctree modeling}

\item{...}{additional arguments to be passed to partykit::ctree}
}
\description{
This function trains a partykit::ctree using data from a passed data.frame.
}

