% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-tree.R
\name{ctree.subscore}
\alias{ctree.subscore}
\title{fit a ctree with passed data and compute and return its spread score}
\usage{
ctree.subscore(i, df, nSample = NULL, ctl = NULL, responseVar, ignoreCols)
}
\arguments{
\item{df}{data.frame with response variable to be explained and all explanatory variables to be used in ctree classificaiton}

\item{nSample}{number of samples if hte spread score should be run on a subset of the data}

\item{ctl}{ctree control to specify ctree model parameters}

\item{responseVar}{the name of the variable that the ctree is trying to explain}

\item{ignoreCols}{an optional list of columns in df, but that should be excluded from the ctree modeling}
}
\description{
This function trains a ctree using data from a passed data.frame and returns the resulting spread score.
}

