% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-dataSource.R
\name{DataSource}
\alias{DataSource}
\title{Main S3 class that defines the data source interface}
\usage{
DataSource()
}
\description{
The functions defined in the DataSource class must be implemented in the custom data source for each 
new data set, taking steps to convert data and normalize data access. It is good practice to ensure that
custom data sources "extend" DataSource by using the \code{DataSource()} constructor as the line that creates the
custom DataSource S3 objects, and then overriding functions as required.
}
\details{
\code{DataSource} Functions as a sort of interface definition and/or default implementation for 
all data source functions. It is good practice to have custom data sources extend this one because
it provides default implementations for all functions and a handfull of useful utility functions.
}
\seealso{
\code{\link{TestData}}, \code{\link{sanityCheckDataSource}}
}

