% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{datesToEpoch}
\alias{datesToEpoch}
\title{Convert POSIXct data frame columns to integer seconds since the epoch}
\usage{
datesToEpoch(df)
}
\arguments{
\item{df}{Data frame of feature data}
}
\value{
A data frame identical to the one passed in, but with integer columns replacing POSIXct ones
}
\description{
Searches the data frame for columns with 'POSIXct' in their class values and 
converts them to integers representing seconds since the epoch
}

