% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-dbUtil.R
\name{dbCfg}
\alias{dbCfg}
\title{Parse standard database configuration file format}
\usage{
dbCfg(filePath)
}
\arguments{
\item{filePath}{The file path to the config file to be read in. Must be absolute or relative to the working directory from \code{getwd()}}
}
\description{
Read database config file with key=value format:
\code{key=value
key2=value2}
etc. into a named list.
Keys must include \code{dbType} passed in to \code{dbDriver} to load the appropriate driver (i.e. MySQL, PostgreSQL, etc.)
and any parameters like \code{host}, \code{user}, \code{password}, \code{dbname} that should be passed into the dbConnect function call for your driver.
}
\details{
The result of this function call is just a list with named entries. As long as the entries are valid, it can be generated in other ways.
Note that the ability to specify a single file that contains database configuration information allows for 
multiple users to share a code based, but provide their personal credentials in their copy of the relevant cfg file.
Due to the inevitability of code conflicts with these files where user names and passwords are different for each user
and the fact that it is insecure and unwise to include database credentials in any version control system, this means 
that such dbCfg files should not be checked into version control.
}
\seealso{
\code{\link{conf.dbCon}}, \code{\link{run.query}}
}

