% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetDocs.R
\docType{data}
\name{ERLE_ZIP_LOCATION}
\alias{ERLE_ZIP_LOCATION}
\title{Zipcode to location mapping}
\format{A data frame with 43191 rows and 7 variables:
\describe{
  \item{zip}{zip code, as a number}
  \item{city}{city the zip code is in}
  \item{state}{two letter abbreviation for the state the zip code is in}
  \item{latitude}{latitude of the zip code center (post office?)}
  \item{longitude}{longitude of the zip code center (post office?)}
  \item{timezone}{timesone, as offste relative to GMT time}
  \item{dst}{indicator for participation in daylight savings time}
}}
\source{
\url{http://can.not.remember/}
}
\usage{
data(ERLE_ZIP_LOCATION)
}
\description{
A dataset containing columns with US zip codes and the corresponding locaiton information
}
\keyword{datasets}

