% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{exportFeatureAndShapeResults}
\alias{exportFeatureAndShapeResults}
\title{Export feature run and load shape results}
\usage{
exportFeatureAndShapeResults(feature.rdata, shape.results.rdata = NULL,
  format = "hdf5", prefix = "")
}
\arguments{
\item{feature.rdata}{File path to an RData file with feature data}

\item{shape.results.rdata}{Optional file path to an RData file containing load shape clustering results}

\item{format}{Export data format - one of the ones supported by exportData()}

\item{prefix}{Optional prefix to put n froun of all feature names}
}
\description{
Loads feature data and load shape clustering data from RData files and 
saves them into the selected export format
}

