% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{fixNames}
\alias{fixNames}
\title{Prepare data frame column names for export}
\usage{
fixNames(df, prefix = "")
}
\arguments{
\item{df}{The data frame whose columns are to be renamed}

\item{prefix}{An optional prefix to place in front of all column names}
}
\value{
A data frame identical to the one passed in, but with new column names.
}
\description{
Removes punctuation from data frame column names, replacing all with underscores
and removing underscores that are repeated one after another
}

