% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-tree.R
\name{groupHist}
\alias{groupHist}
\title{ctree classificaiton results plot for a binary response variable}
\usage{
groupHist(df, df.ct, responseCol, title = "Group probabilities",
  idCol = "id", xlab = "Enrollment probability (\%)", normalize = FALSE,
  compareData = NULL, colors = NULL)
}
\arguments{
\item{df}{data.frame with response variable to be explained and all explanatory variables to be used in ctree classificaiton}

\item{df.ct}{ctree already trained on data, that is used to classify the data points in each sub sample.}

\item{responseCol}{the name of the variable that the ctree is trying to explain}

\item{title}{figure title of the plot}

\item{idCol}{column with data row identifiers in it, which is removed from the df}

\item{xlab}{label for x-axis of the figure}

\item{normalize}{boolean for whether the x-axis should be percentages compared to the sample mean or absolute percentages}

\item{compareData}{optional data.fram of one additional set of bar values to be plotted as well.}

\item{colors}{optional named list specifying the figure colors.}
}
\description{
plot a histogram of customer segments and enrollment percentages, based on the membership of ctree leaf nodes.
The plot is a "histogram" whose x-axis is the average "True" response for each ctree leaf node and whose 
height is the number of customers in each corresponding group. Strong results have nodes with high membership
far from the sample mean.
}

