% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator.callAll}
\alias{iterator.callAll}
\title{Run multiple feature algorithms specified in a list of functions}
\usage{
iterator.callAll(meterDataClass, ctx, fnList, ...)
}
\arguments{
\item{meterDataClass}{The \code{\link{MeterDataClass}} object that contains the data to be analyzed.}

\item{ctx}{The ctx environment that configures feature runs and provides a place to store and pass data across feature function calls.}

\item{fnList}{The list of feature extraction functions to be run against the \code{\link{MeterDataClass}}.}

\item{...}{Additional arguments that will be passed into the feature functions.}
}
\description{
This function passes the meterDataClass object into
every feature function in the list found in the parameter \code{fnList}, 
concatenating all the results into a single list with named values.
}

