% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator.callAllFromCtx}
\alias{iterator.callAllFromCtx}
\title{Run feature algorithms specified in the ctx environment}
\usage{
iterator.callAllFromCtx(meterId, ctx, ...)
}
\arguments{
\item{meterId}{The id that the data source can use to lookup the meter data of interest.}

\item{ctx}{The ctx environment that configures feature runs and provides a place to store and pass data across feature function calls.}

\item{...}{Additional arguments that will be passed into the feature functions.}
}
\description{
This function loads the meter data associated with the passed \code{meterId} and creates 
a MeterDataClass instance using that data. It then passes that meter data object into
every feature function in the list found in the ctx environment under \code{fnVector}, 
concatenating all the results into a single list with named values.
}

