% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator.iterateMeters}
\alias{iterator.iterateMeters}
\title{Run features for a list of meter ids}
\usage{
iterator.iterateMeters(meterList, custFn, ctx, ...)
}
\arguments{
\item{meterList}{The list of meterIds to use to instantiate a \code{MeterDataClass} objects}

\item{custFn}{The feature extraction function to run on each instance of \code{MeterDataClass}.}

\item{ctx}{The ctx environment for the feature extraction.}

\item{...}{Additional arguments to be passed to the feature extraction function.}
}
\description{
Utility function that runs the configured set of feature functions on the data from a list of passed 
meter ids. This is useful for scripting feature extraction on an arbitrary number of meters.
}

