% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator.runMeter}
\alias{iterator.runMeter}
\title{Run features for a single meterId}
\usage{
iterator.runMeter(meterId, custFn, ctx, ...)
}
\arguments{
\item{meterId}{The meterId to use to instantiate a \code{MeterDataClass} object}

\item{custFn}{The feature extraction function to run on the meter data.}

\item{ctx}{The ctx environment for the feature extraction.}

\item{...}{Additional arguments to be passed to the feature extraction function.}
}
\description{
Utility function that runs the configured set of feature functions on a single passed 
meter. This is useful for testing and feature development, but also as the function 
called by parallelizable methods like apply, and the *ply function of plyr to run feature
extraction in on multiple cores of a computer.
}

