% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator.todf}
\alias{iterator.todf}
\title{Flatten feature data returned by iterator functions into a \code{data.frame}.}
\usage{
iterator.todf(lofl)
}
\arguments{
\item{lofl}{The list of lists for feature values that should be flattened into a single data frame, with one row per meterId.}
}
\description{
Itertor functions like \code{iterator.iterateMeters} return lists of lists of derived features indexed by meterIds. 
This function flattens the scalar values in these lists into data.frame, with one row per meterId and columns for every named feature found.
}
\details{
This function ignores non-scalars, so diagnostic data and other complex data structures can be returned by feature algorithms without 
interfering with the task of creating clean vectors of features for each meter. The columns of the returned data.frame are
a super set of all the features returned for every MeterDataClass with computed features. Missing features are given values of NA.
}

