% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{mergeShapeFeatures}
\alias{mergeShapeFeatures}
\title{Merge load shape features into feature data frame}
\usage{
mergeShapeFeatures(features, shape.results)
}
\arguments{
\item{features}{Data frame of feature data}

\item{shape.results}{Load shape clustering and assignment results object to pull features from}
}
\value{
A data frame identical to the one passed in, but with new laod shape feature columns
}
\description{
Pulls load shape features from a shape results object and appends them to 
an existing feature data frame
}

