% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-customer.R
\name{MeterDataClass}
\alias{MeterDataClass}
\title{Main S3 class that holds and normalizes VISDOM customer meter data}
\usage{
MeterDataClass(id, geocode = NULL, weather = NULL, data = NULL,
  gasData = NULL, useCache = T, doSG = F, rawData = NULL)
}
\arguments{
\item{id}{Unique identifier used by data source to identify a customer meter.
Even with numerical id's in a database, it is a good idea to make these
character strings in R.}

\item{geocode}{Primary geographic locator used by the data source. This is
very often a zip code, but data sources can implement it as census block
or any geographic category that covers all meters, with each meter
associated with only one geocode. If it is not provided, the data source
is used to query for the meter's location.}

\item{weather}{Reference to instance of \link{WeatherData} object that provides
weather data for the meter's location. The DATA_SOURCE is used to try to
load this information for the relevant geocode if it is not provided.}

\item{data}{Tabluar format dataframe of electric meter data for the meter identified
by \code{id}, with each row covering all observations for a meter for a day, with
columns including at least: id, date, and hourly or 15 minute readings.}

\item{gasData}{Tabluar format dataframe of gas meter data, with each row covering all
observations for a meter for a day, with columns including at least: id, date
and daily readings. If this is not provided, the gas meter data is
loaded using the data source if the data source supports gas data.
\code{gasData} is typically passed in as an optimmization
to avoid unnecessary database queries during bulk meter feature extraction.}

\item{useCache}{Data caching option indicating to the data source whether cached
data should be used to populate the class. Data sources implement their own
caching, but the expectation is that they will hit the underlying data base
once to get the data in the first place and then save it to a local cache from
which it will be retrieved on subsequent calls. See \link{run.query} for details.}

\item{doSG}{Option indicating whether expensive solar geometry calculations
that rely on the \link{solaR} package should be performed. Off by default for
performance reasons. See \link{solarGeom} for details.}

\item{rawData}{Tabluar format dataframe of electric meter data, with each row covering all
observations for a meter for a day, with columns including at least: id, date,
and hourly or 15 minute readings. If this is not provided and \code{data} is not provided,
the meter data is loaded using the data source. \code{rawData} is typically passed in as
an optimization to avoid unnecessary database queries during bulk meter feature
extraction.}
}
\description{
Data structures sand functions used to store, manipulate, and visualize customer
meter data. This class provides structure and standardization to the representation
of meter data in VISDOM, with support for passing in pre-loaded version of data 
that is otherwise expensive to load individually. It therefore enables feature 
functions, figures, etc. to all be written to draw upon standard structures and
functions and centralizes work on performance.
}
\details{
MeterDataClass is often called within the iterator framework or using DATA_SOURCE$getMEterDataClass(id) for a data source.
}
\examples{
\dontrun{
DATA_SOURCE = TestData(10)
cust = MeterDataClass(id='14')
plot(cust)
}
}
\seealso{
\code{\link{WeatherClass}}
}

