% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-base.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}}
}}
\keyword{internal}

