% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-regression.R
\name{regressor.piecewise}
\alias{regressor.piecewise}
\title{Break a vector into piecewise regressor columns, using the specified breaks}
\usage{
regressor.piecewise(regressor, bins, diverge = F)
}
\arguments{
\item{regressor}{The single array of regressor values to be broken out into piecewise columns.}

\item{bins}{- numerical arry whose values are the break points defining the piecewise regressors}

\item{diverge}{- Defautl FALSE: Whether the first column contains the distance of
the value from the bottom change point (rather than from 0)}
}
\description{
break a vector out for continuous piecewise regression (i.e. the fitted
segments will join eachother at each end) into a matrix whose row
totals are the original values, but whose columns divide the value across
a set of bins, so 82 across bins with boundaries c(50,60,65,70,80,90)
becomes the row 50,10,5,5,10,2,0, which sums to 82...
This is very useful for finding rough change points in thermal response
as is expected for buildings with clear setpoints
}
\details{
TODO: This can create a column of zeros, which should break the regression
so we might need to prune the columns when we're done and keep track of
which bins are in play when comparing across regressions
}
\seealso{
\code{\link{piecewise.regressor}} for 'apply' syntax
}

