% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-regression.R
\name{regressor.split}
\alias{regressor.split}
\title{Split a single column of values into multiple columns as determined by group membership}
\usage{
regressor.split(regressor, membership = NULL)
}
\arguments{
\item{regressor}{The vector of data to be split}

\item{membership}{A vector of the same length as the regressor that assigns group membership categorically.
If the regressor is a factor, no membership assignment is required.}
}
\value{
A matrix of length(sort(unique(membership))) columns, that is all
zeros except for the regressor values that match the membership values
corresponding to the column. This supports regression with separate
coefficints for each group defined by the membership
for example, regressor.split(Tout,dates$hour) will return a matrix with 24
columns, where the only non-zero entry per row contains the Tout value in
the column corresponding to the hour of day it was recorded
}
\description{
Split a column of data by membership in discrete groups (as with factor levels) 
data into multiple columns containing just the values corresponding to the membership
indicator associated with the column and zeros otherwise.
}

