% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-regression.R
\name{regressorDF}
\alias{regressorDF}
\title{data.frame formatting of \code{\link{MeterDataClass}} data}
\usage{
regressorDF(meterData, norm = FALSE, rm.na = FALSE)
}
\arguments{
\item{meterData}{The meter data class to be converted}
}
\description{
Given a MeterDataClass instance (meterData), regressorDF returns a data.frame consisting of a standard set of
regressor columns suitable for passing into a call to lm.
}

