% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-dbUtil.R
\name{showCons}
\alias{showCons}
\title{List all open database connections}
\usage{
showCons(cfg)
}
\arguments{
\item{cfg}{parsed connection configuration data}
}
\description{
Use this to see whether too many connections are open
}

