% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-tree.R
\name{spread.boot}
\alias{spread.boot}
\title{bootstrap spread scores across several runs of the ctree algorithm}
\usage{
spread.boot(df, df.ct, nRun = 100, nSample = NULL, responseVar, ignoreCols)
}
\arguments{
\item{df}{data.frame with response variable to be explained and all explanatory variables to be used in ctree classificaiton}

\item{df.ct}{ctree already trained on data, that is used to classify the data points in each sub sample.}

\item{nRun}{number of bootstrapping runs to perform}

\item{nSample}{number of samples (with replacement) for each run}

\item{responseVar}{the name of the variable that the ctree is trying to explain}

\item{ignoreCols}{an optional list of columns in df, but that should be excluded from the ctree modeling}
}
\description{
This function repeatedly randomly sub-samples the passed data.frame to provide points for classificaiton 
by the passed ctree model and calculates the resulting spread score for classifications of each subset 
and returns a vector of the resulting scores. It is useful for determining the robustness of the spread score 
for a particular data set.
}

