% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDataSource.R
\name{TestData}
\alias{TestData}
\title{Implementation of a data source that generates synthetic data for testing and examples}
\usage{
TestData(n = 100)
}
\arguments{
\item{n}{Number of meters for the data source to provide synthetic data for: determines 
the return values for functions like \code{DATA_SOURCE$getIds()} and and \code{DATA_SOURCE$getAllData()}}
}
\description{
Example implementation of the data source functions as a coherent data source. This one 
simply generates random synthetic data that conform to the formats required by \code{\link{MeterDataClass}}
and \code{\link{WeatherClass}}, returning valid, but meaningless data for all calls.
}
\details{
\code{TestData} is instantiated with a certain number of meters to generate data for. It then generates
random data for that number and consumes that data to support the rest of its functions. This class
can be used for examples, learning how data sources work, and providing data for tests of feature algorithms
and meter data related classes.
}
\examples{
\dontrun{
DATA_SOURCE = TestData(n=100)
idList = DATA_SOURCE$getIds()
dataMatrix = DATA_SOURCE$getAllData()
}
}
\seealso{
\code{\link{DataSource}}
}

