% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-weather.R
\name{WeatherClass}
\alias{WeatherClass}
\title{S3 class that holds and normalizes weather data}
\usage{
WeatherClass(geocode, raw = NULL, doMeans = T, useCache = F, doSG = F)
}
\arguments{
\item{geocode}{Primary geographic locator used by the data source. This is
very often a zip code, but data sources can implement it as census block
or any geographic category that covers all meters, with time series
observations of weather data associated with a geocode.}

\item{doMeans}{Indicator of whether to calculate and retain daily and monthly
mean values of observations. Disable when not in use for faster performance.}

\item{useCache}{Data caching option indicating to the data source whether cached
data should be used to populate the class. Data sources implement their own
caching, but the expectation is that they will hit the underlying data base
once to get the data in the first place and then save it to a local cache from
which it will be retrieved on subsequent calls. See \link{run.query} for details.}

\item{doSG}{Option indicating whether expensive solar geometry calculations
that rely on the \link{solaR} package should be performed. Off by default for
performance reasons. See \link{solarGeom} for details.}
}
\description{
Standard format and functions for loading, manipulating, and visualizing weather data in VISDOM.
}
\details{
\code{WeatherData} is compatible by default with the output (i.e. database table) of
  python weather data scraping code that draws upon NOAA's NOAA Quality Controlled 
  Local Climatological Data (QCLCD) ftp server. See \code{http://github.com/sborgeson/local-weather}
  for code to download and convert weathre data into CSV files, the weather data zip files at
  \code{http://www.ncdc.noaa.gov/orders/qclcd/}, and the QCLCD homepage here
  \code{https://www.ncdc.noaa.gov/data-access/land-based-station-data/land-based-datasets/quality-controlled-local-climatological-data-qclcd}
}
\section{Parameters}{

}
\examples{
\dontrun{
DATA_SOURCE = TestData(10)
weather = WeatherData(geocode='12601')
plot(weather)
}
}
\seealso{
\code{\link{MeterDataClass}}
}

