% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{writeH5Data}
\alias{writeH5Data}
\title{Write feature data frame to an hdf5 file}
\usage{
writeH5Data(data, fName, label, filePath = NA)
}
\arguments{
\item{data}{The feature data frame to be written}

\item{fName}{The name of the hdf5 formatted file to write the data to}

\item{label}{The name of the data table within the hdf5 file}
}
\description{
Write feature data frame to an hdf5 file
}

