% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetDocs.R
\docType{data}
\name{ZIP_TO_ZCTA}
\alias{ZIP_TO_ZCTA}
\title{A zip code to census zip code tabulation area mapping}
\format{A data frame with 41979 rows and 5 variables:
\describe{
  \item{ZIP}{zip code, as a 0 padded string}
  \item{ZIPType}{zip code type}
  \item{CityName}{Name of the city the zip code is in}
  \item{StateAbbr}{Two letter abbreviation for the state the zip code is in}
  \item{ZCTA}{the census zip code tabulation area (ZCTA) the zip code most overlaps with}
}}
\source{
\url{http://some.random.helpful.blog/}
}
\usage{
data(ZIP_TO_ZCTA)
}
\description{
A dataset containing columns with US zip codes and the corresponding ZCTA ids
}
\keyword{datasets}

