% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{applyDateFilters}
\alias{applyDateFilters}
\title{MOY - list of months of the year to include, using 1 for Jan through 12 for Dec
DOW - days of week to include, using 1 for Sun and 7 for Sat
start.date - the first day of data to include: all dates before this date are excluded
end.date - the last day of data to include: all dates after this date are excluded}
\usage{
applyDateFilters(df, filterRules = NULL, dateCol = "dates")
}
\description{
MOY - list of months of the year to include, using 1 for Jan through 12 for Dec
DOW - days of week to include, using 1 for Sun and 7 for Sat
start.date - the first day of data to include: all dates before this date are excluded
end.date - the last day of data to include: all dates after this date are excluded
}

