% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{exportFeatureAndShapeResults}
\alias{exportFeatureAndShapeResults}
\title{Export feature run and load shape results}
\usage{
exportFeatureAndShapeResults(feature.data, shape.results.data = NULL,
  format = "hdf5", prefix = "")
}
\arguments{
\item{feature.data}{File path to an RData file with feature data data frame or the data frame itself}

\item{shape.results.data}{Optional file path to an RData file containing load shape clustering results 
or the results object itself. i.e. results from 
\code{visdomloadshape::shapeFeatures(visdomloadshape::shapeCategoryEncoding())}}

\item{format}{Export data format - one of the ones supported by exportData()}

\item{prefix}{Optional prefix to put n froun of all feature names}
}
\description{
Loads feature data and load shape clustering data from RData files and 
saves them into the selected export format
}

