% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{exportShapes}
\alias{exportShapes}
\title{Save load shape results}
\usage{
exportShapes(shape.results, prefix = "", format = "hdf5")
}
\description{
Exports standardized load shape clustering and assignment data into a 
corresponding set of exported data tables
}

