% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator.iterateZip}
\alias{iterator.iterateZip}
\title{Iterate over zip codes, extracting features for meters in each}
\usage{
iterator.iterateZip(zipList, custFn, cacheResults = F, ctx = NULL,
  clearCachedResultsFromRAM = F, ...)
}
\arguments{
\item{zipList}{List of all zip codes to iterate over}

\item{custFn}{The feature function to call on \code{MeterDataClass} instances from within each zip code}

\item{cacheResults}{A boolean flag that indicates whether feature results should be cached as RData. 
If true, the cached data will be looked up prior to feature extraction to bypass running 
features for the given zip code while returning the cached results. This can prevent duplicate processing
and allow an interrupted batch process to resume processing where it left off.
The cache directory is `getwd()` by default, but can be overridden using `ctx$resultsCache`.}

\item{ctx}{The ctx environment that configures the feature run.}

\item{clearCachedResultsFromRAM}{A boolean that instructs the code to drop previous results from RAM if caching 
is in effect. This allows for larer runs to be executed and cached without accumulating in RAM. The idea is 
that after the run, this function will be called again, with useCache=TRUE and this flag set to FALSE, which 
will simply load and concatenate results from disk to re-create the full set.}

\item{...}{Arguments to be passed into the feature function(s).}

\item{zipList}{List of zip codes over which to run the feature.}

\item{custFn}{The feature function(s) to call on \code{MeterDataClass} instances from within the zip code}

\item{cacheResults}{A boolean flag that indicates whether results should be cached as RData. 
If true, the cached data will be looked up prior to feature extraction to bypass running 
features for the given zip code while returning the cached results. This can prevent duplicate processing
and allow an interrupted batch process to resume processing where it left off.
The cache directory is `getwd()` by default, but can be overridden using `ctx$resultsCache`.}
}
\description{
Function that iterates through all passed zip codes, looks up local weather data (once) and 
looks up the list of meter ids for each and calls iterator.iterateMeters with those meter 
ids and the per-loaded weather data. This runs faster than calling ids individually, which 
can load similar weather data over and over.
}

