% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{runDateFilterIfNeeded}
\alias{runDateFilterIfNeeded}
\title{Run date filters on raw meter data}
\usage{
runDateFilterIfNeeded(ctx)
}
\arguments{
\item{ctx}{Context object that is an R environment with named parameters}
}
\description{
Runs \code{\link{applyDateFilters}} if the appropriate values are found in the ctx and the data is not yet
flagged as filtered.
}

