% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{writeCSVData}
\alias{writeCSVData}
\title{Write feature data frame to a csv file}
\usage{
writeCSVData(data, fName, label = NA, filePath = NA)
}
\arguments{
\item{data}{The feature data frame to be written}

\item{fName}{The name of the csv file to write the data to}

\item{label}{Unused, but present for compatibility with other write* fucntions}

\item{filePath}{optional path to the location where exported files should be written (if applicable). Default is \code{getwd()}}
}
\description{
Write feature data frame to a csv file
}

