% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-export.R
\name{writeDatabaseData}
\alias{writeDatabaseData}
\title{Write feature data frame to a database}
\usage{
writeDatabaseData(data, tableName, label = NULL, conn, overwrite = TRUE)
}
\arguments{
\item{data}{The feature data frame to be written}

\item{tableName}{The name of the table to write the data to}

\item{label}{Unused, but present for compatibility with other write* fucntions}

\item{conn}{A DBI dbConnection object to the database that will host the table}

\item{overwrite}{Boolean indicator for whether the data written should overwrite any existing table or append it}
}
\description{
Write feature data frame to a database
}

